core/build.rs: set user_version PRAGMA in generated schema.sql
The generated schema did not include the `user_version` which tracks
which migration is the latest one. This made the README.md example of
creating a database manually fail because it would be initialized with a
`user_version` of 0 and then `mailpot` would attempt to apply migrations
to it.

```shell
$ sqlite3 /path/to/db < ./core/src/schema.sql
$ cargo run --bin mpot -- [some command]
ERROR - 1 no such table: templates in "
ALTER TABLE templates RENAME TO template;"
[1] Error returned from sqlite3 no such table: templates.
```

Fixes #1

https://git.meli.delivery/meli/mailpot/issues/1

Signed-off-by: Manos Pitsidianakis <manos@pitsidianak.is>